requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
idna==3.10
jieba3k==0.35.1
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
pillow==11.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
